more fixups

+++ pandas/tests/interchange/test_impl.py
@@ -101,9 +101,9 @@ def test_dataframe(data):
 def test_missing_from_masked():
     df = pd.DataFrame(
         {
-            "x": np.array([1, 2, 3, 4, 0]),
+            "x": np.array([1, 2, 3, 4, 0], dtype=float),
             "y": np.array([1.5, 2.5, 3.5, 4.5, 0]),
-            "z": np.array([True, False, True, True, True]),
+            "z": np.array([True, False, True, True, True], dtype=float),
         }
     )
